fix: structure
