+++ src/features/Auth/components/index.tsx
@@ -1,3 +1,3 @@
 export { AuthButtons } from './button';
 export { AuthInputs } from './inputs';
-export { FormTabs, Tab } from './tabs';
+export { Tab } from './tabs';
