+++ src/features/Auth/components/tabs/index.module.scss
