--- src/features/Auth/form/components/button/types.ts
@@ -1,8 +1,0 @@
-export interface IAuthButtonsProps {
-  buttonProps: {
-    authType: boolean;
-    code: boolean;
-    isValid: boolean;
-    handleAuthTypeChange(): void;
-  };
-}
