+++ src/features/Auth/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/styles/mixins/gradient.scss' as *;
+@use '../../assets/styles/mixins/gradient.scss' as *;
 
 .auth {
   display: flex;
